Feature/add projects (#2)
This adds the web app I have been working on, with the commands to run
it working. It also adds the basic Rust servers for the backend. These
have no real functionality, and were generated mostly with the help of AI.

It also adds different infrastructure systems, like npm/cargo workspaces and containerization.